+++ packages/implementations/react/src/input/Select/src/select.spec.tsx
@@ -21,7 +21,7 @@ const mockedOptions = [
   },
 ];
 
-const mockedValues = ["option-1"];
+const mockedValues = [{ label: "Option 1", value: "option-1" }];
 
 const makeSut = (
   rest?: Partial<Omit<SelectProps, "id" | "name">>,
@@ -95,7 +95,7 @@ describe("GIVEN <Select />", () => {
     });
     fireEvent.click(screen.getByTestId("select-element"));
     fireEvent.click(screen.getByLabelText(mockedOptions[0].label));
-    expect(mockedOnChange).toHaveBeenCalledWith([mockedOptions[0].value]);
+    expect(mockedOnChange).toHaveBeenCalledWith([mockedOptions[0]]);
     fireEvent.click(screen.getByLabelText(mockedOptions[0].label));
     expect(mockedOnChange).toHaveBeenCalledWith([]);
   });
@@ -108,7 +108,7 @@ describe("GIVEN <Select />", () => {
     });
     expect(
       screen.getByTestId<HTMLSelectElement>(
-        `option-selected-${mockedValues[0]}`,
+        `option-selected-${mockedValues[0].value}`,
       ),
     ).toBeDefined();
   });
@@ -123,7 +123,7 @@ describe("GIVEN <Select />", () => {
     });
     fireEvent.click(screen.getByTestId("select-element"));
     fireEvent.click(screen.getByText(mockedOptions[0].label));
-    expect(mockedOnChange).toHaveBeenCalledWith([mockedOptions[0].value]);
+    expect(mockedOnChange).toHaveBeenCalledWith([mockedOptions[0]]);
     fireEvent.click(
       screen.getByTestId(
         `option-selected-${mockedOptions[0].value}-button-close`,